+++ src/components/EditPopup.tsx
@@ -1,5 +1,9 @@
-"use client"
+'use client';
+
 import * as React from 'react';
+
+import { makeApiCall } from '@/utils/ApiRequest';
+import EditNoteIcon from '@mui/icons-material/EditNote';
 import Button from '@mui/material/Button';
 import Dialog from '@mui/material/Dialog';
 import DialogActions from '@mui/material/DialogActions';
@@ -7,22 +11,44 @@ import DialogContent from '@mui/material/DialogContent';
 import DialogTitle from '@mui/material/DialogTitle';
 import Slide from '@mui/material/Slide';
 import TextField from '@mui/material/TextField';
-import axios from 'axios';
+import { TransitionProps } from '@mui/material/transitions';
 
-const Transition = React.forwardRef(function Transition(props, ref) {
+const Transition = React.forwardRef<
+  unknown,
+  TransitionProps & { children: React.ReactElement }
+>(function Transition(props, ref) {
   return <Slide direction="up" ref={ref} {...props} />;
 });
 
-const BaseURL = process.env.BaseURL;
-export default function PopUpSlide({setTableData}) {
+interface TableItem {
+  title: string;
+  start: string;
+  end: string;
+  note: string;
+}
+
+interface TimeTableRecord {
+  date: string;
+  id: string;
+  is_finished: boolean;
+  table_item: TableItem[];
+  total_min: number;
+  user_id: string;
+}
+
+interface EditPopupProps {
+  rowNumber: number;
+  previousData: TableItem;
+  setTableData: React.Dispatch<React.SetStateAction<TimeTableRecord | null>>;
+}
+
+const EditPopup: React.FC<EditPopupProps> = ({
+  rowNumber,
+  previousData,
+  setTableData,
+}) => {
   const [open, setOpen] = React.useState(false);
-  const [formData, setFormData] = React.useState({
-    name: '',
-    start_time: '',
-    end_time: '',
-    duration: '',
-    description: '',
-  });
+  const [formData, setFormData] = React.useState(previousData);
 
   const handleClickOpen = () => {
     setOpen(true);
@@ -31,42 +57,53 @@ export default function PopUpSlide({setTableData}) {
   const handleClose = () => {
     setOpen(false);
   };
-  const handleChange = (e) => {
+
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const { name, value } = e.target;
     setFormData({ ...formData, [name]: value });
   };
 
-  const handleSubmit = async() => {
+  const handleSubmit = async () => {
     try {
       const Payload = {
         data: {
           type: 'time-table',
           attributes: {
-            title: formData.name,
-            start: formData.start_time,
-            end: formData.end_time,
-            note: formData.description
-          }
-        }
-      }
-      const res = await axios.patch(BaseURL + "time-table", Payload);
-      setTableData(res.data.data.attributes)
-      setOpen(false)
-      return res;
+            row: rowNumber,
+            title: formData.title,
+            start: formData.start,
+            end: formData.end,
+            note: formData.note,
+          },
+        },
+      };
+      const result = await makeApiCall({
+        path: `subject/time-table`,
+        method: 'PUT',
+        payload: Payload,
+      });
+
+      setTableData(result.data.attributes);
+      setOpen(false);
+      return result;
     } catch (error) {
-      console.log(error)
+      console.log(error);
     }
-   
+
     handleClose();
   };
 
   return (
     <React.Fragment>
-      <div style={{width: "100%", display: "flex", alignItems: 'center', justifyContent: "center"}}>
-      <Button variant="outlined" onClick={handleClickOpen}>
-        {/* {editingData ? 'Edit' : 'Push I'} */}
-        Push Item
-      </Button>
+      <div
+        style={{
+          width: '100%',
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'center',
+        }}
+      >
+        <EditNoteIcon sx={{ color: 'red' }} onClick={handleClickOpen} />
       </div>
       <Dialog
         open={open}
@@ -75,7 +112,7 @@ export default function PopUpSlide({setTableData}) {
         onClose={handleClose}
         aria-describedby="alert-dialog-slide-description"
       >
-        <DialogTitle>Create a new entry</DialogTitle>
+        <DialogTitle>Edit entry</DialogTitle>
         <DialogContent>
           <TextField
             autoFocus
@@ -84,8 +121,8 @@ export default function PopUpSlide({setTableData}) {
             label="Title"
             type="text"
             fullWidth
-            name="name"
-            value={formData.name}
+            name="title"
+            value={formData.title}
             onChange={handleChange}
             required
           />
@@ -95,8 +132,8 @@ export default function PopUpSlide({setTableData}) {
             label="Start Time"
             type="time"
             fullWidth
-            name="start_time"
-            value={formData.start_time}
+            name="start"
+            value={formData.start}
             onChange={handleChange}
             required
           />
@@ -106,8 +143,8 @@ export default function PopUpSlide({setTableData}) {
             label="End Time"
             type="time"
             fullWidth
-            name="end_time"
-            value={formData.end_time}
+            name="end"
+            value={formData.end}
             onChange={handleChange}
           />
           <TextField
@@ -116,17 +153,18 @@ export default function PopUpSlide({setTableData}) {
             label="Description"
             type="text"
             fullWidth
-            name="description"
-            value={formData.description}
+            name="note"
+            value={formData.note}
             onChange={handleChange}
           />
         </DialogContent>
         <DialogActions>
           <Button onClick={handleClose}>Cancel</Button>
-          <Button onClick={handleSubmit}>Add</Button>
+          <Button onClick={handleSubmit}>{'Save'}</Button>
         </DialogActions>
       </Dialog>
     </React.Fragment>
   );
-}
+};
 
+export default EditPopup;
